image-builder: move subcommand setup into functions
Create separate functions for each subcommand and a top-level one for
creating the root command.  The setupRootCmd() function sets up the root
command and calls the setupX() function for each of its subcommands.

This is a very simple code-shuffling commit.  It's meant to make
navigating the CLI creation easier by localising the creation and
configuration of each subcommand.
